fix(reload firefox): check if firefox is running before killing it

+++ dotfiles/config/hypr/scripts/reload_firefox.sh
@@ -2,5 +2,8 @@
 
 position=$(hyprctl clients -j | jq 'map(select(.class=="firefox"))[].workspace.id')
 
-pkill firefox
-hyprctl dispatch exec [workspace $position silent] firefox > /dev/null 2>&1
+if pgrep -x firefox > /dev/null
+then
+    pkill firefox
+    hyprctl dispatch exec [workspace $position silent] firefox > /dev/null 2>&1
+fi
